Feature/session

+++ src/main/java/spring/board/service/BoardService.java
@@ -6,9 +6,14 @@ import spring.board.dao.JdbcTemplate.JdbcCommentDao;
 import spring.board.dao.JdbcTemplate.JdbcFileDao;
 import spring.board.dao.MyBatis.BoardMapper;
 import spring.board.dao.MyBatis.FileMapper;
-import spring.board.dto.*;
-import spring.board.entity.Board;
-import spring.board.entity.Paging;
+import spring.board.domain.Board;
+import spring.board.domain.Paging;
+import spring.board.dto.board.*;
+import spring.board.dto.common.PageInfo;
+import spring.board.dto.file.FileRequest;
+import spring.board.dto.file.FileResponse;
+import spring.board.dto.user.UserSession;
+import spring.board.util.SessionUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -19,7 +24,6 @@ import java.util.*;
 @Service
 //@RequiredArgsConstructor
 public class BoardService {
-    private final UserService userService;
     private final FileService fileService;
     private final CommentService commentService;
     private final BoardMapper boardMapper;
@@ -28,8 +32,7 @@ public class BoardService {
     //    private final JdbcBoardDao boardDao;
 //    private final JdbcFileDao fileDao;
 
-    public BoardService(UserService userService, FileService fileService, CommentService commentService, JdbcBoardDao boardDao, JdbcFileDao fileDao, JdbcCommentDao commentDao, BoardMapper boardMapper, FileMapper fileMapper) {
-        this.userService = userService;
+    public BoardService(FileService fileService, CommentService commentService, BoardMapper boardMapper, FileMapper fileMapper) {
         this.fileService = fileService;
         this.commentService = commentService;
         this.fileMapper = fileMapper;
@@ -41,47 +44,49 @@ public class BoardService {
 
 
 
-    public String post(BoardRequest boardRequest, HttpServletRequest request) throws IOException {
-        HttpSession session = request.getSession();
-        UserSession userSession = (UserSession) session.getAttribute("USER");
+    public String post(BoardRequest boardRequest, HttpServletRequest request) throws Exception {
+//        HttpSession session = request.getSession();
+//        UserSession userSession = (UserSession) session.getAttribute("USER");
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
         if (userSession != null) {
             Board board = new Board(null, boardRequest.getTitle(), boardRequest.getContent(), userSession.getIdx(), new Timestamp(new Date().getTime()), null);
 //            int bIdx = boardDao.insertPost(board);
-            int bIdx = boardMapper.insertPost(board);
-            if (boardRequest.getFile() != null) {
-                boardRequest.setId(bIdx);
-                fileService.uploadFiletoFtp(boardRequest, request); // 파일 업로드
-                return "게시물 작성 완료";
+            Integer bIdx = boardMapper.insertPost(board);
+
+            if (bIdx == null) {
+                return "게시물 작성 실패";
+            } else {
+                if (boardRequest.getFile() != null) {
+                    boardRequest.setId(bIdx);
+                    fileService.uploadFiletoFtp(boardRequest, request); // 파일 업로드
+                }
             }
         } else {
             System.out.println("로그인을 먼저 해주세요.");
+            return "게시물 작성 실패";
         }
-        return "게시물 작성 실패";
+        return "게시물 작성 완료";
     }
 
     // 페이지
-    public List<BoardResponse> selectAllPosts(int page, int size, int blockSize) {
-        Paging paging = new Paging();
-//        paging.setPaging(page, size, blockSize, boardDao.getTotalCnt());
+    public List<BoardResponse> selectAllPosts(String searchType, String keyword, int curPage, int pageSize, int blockSize) {
+        BoardSearchRequest search = new BoardSearchRequest(searchType, keyword, curPage, pageSize, blockSize, boardMapper.getTotalCnt());
 //        return boardDao.selectPost(paging);
-        paging.setPaging(page, size, blockSize, boardMapper.getTotalCnt());
-        List<BoardResponse> list = boardMapper.selectPost(paging);
+        List<BoardResponse> list = boardMapper.selectPost(search);
         return list;
     }
 
     public PageInfo getPagingInfo(int page, int size, int blockSize) {
-        Paging paging = new Paging();
+        Paging paging = new Paging(page, size, blockSize, boardMapper.getTotalCnt());
 //        paging.setPaging(page, size, blockSize, boardDao.getTotalCnt());
-        paging.setPaging(page, size, blockSize, boardMapper.getTotalCnt());
         PageInfo pageInfo = new PageInfo(paging);
         return pageInfo;
     }
 
     public BoardResponse selectPostByPostId(Integer bIdx) {
-        BoardResponse test = boardMapper.selectPostByPostId(bIdx);
-//        return boardDao.selectPostByPostId(bIdx);
-        return test;
+        BoardResponse boardResponse = boardMapper.selectPostByPostId(bIdx);
+        return boardResponse;
     }
 
     public BoardInfoResponse selectPostsByPostId(Integer bIdx) {
@@ -92,25 +97,25 @@ public class BoardService {
         else fileResponse = new FileResponse(file.getFileNo(), file.getFileName(), file.getPath());
 //        return new BoardInfoResponse(boardDao.selectPostNByPostId(bIdx), fileResponse, commentService.selectCommentsByPostId(bIdx));
 
-        BoardInfoResponse boardInfoResponse = new BoardInfoResponse(boardMapper.selectPostByPostId(bIdx), fileResponse, commentService.selectCommentsByPostId(bIdx));
+        BoardInfoResponse boardInfoResponse =
+                new BoardInfoResponse(
+                        boardMapper.selectPostByPostId(bIdx),
+                        fileResponse,
+                        commentService.selectCommentsByPostId(bIdx));
 
         return boardInfoResponse;
     }
 
-    public List<BoardResponse> selectPostAndCommentByPostId(Integer bIdx) {
-        BoardResponse boardResponse = new BoardResponse();
-        return null;
-    }
-
     // 내 게시물 - 사용자 검색
-    public List<BoardResponse> selectPostsByUserId(int page, int size, int blockSize, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        UserSession userSession = (UserSession) session.getAttribute("USER");
+    public List<BoardResponse> selectPostsByUserId(int page, int size, int blockSize) throws Exception {
+//        HttpSession session = request.getSession();
+//        UserSession userSession = (UserSession) session.getAttribute("USER");
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
         if (userSession != null) {
-            Paging paging = new Paging();
-            paging.setPaging(page, size, blockSize, boardMapper.getTotalCnt());
-            Map<String, Integer> map = new HashMap();
+            Paging paging = new Paging(page, size, blockSize, boardMapper.getTotalCnt());
+//            paging.setPaging(page, size, blockSize, boardMapper.getTotalCnt());
+            Map<String, Integer> map = new HashMap<>();
             map.put("startIndex", paging.getStartIndex());
             map.put("endIndex", paging.getEndIndex());
             map.put("id", userSession.getIdx());
@@ -122,49 +127,57 @@ public class BoardService {
         }
     }
 
-    public String updatePost(Integer bIdx, BoardRequest boardRequest, HttpServletRequest request) {
-        UserSession userSession = userService.getLoginUserInfo(request);
-        Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+    public String updatePost(Integer bIdx, BoardRequest boardRequest) throws Exception {
+//        UserSession userSession = userService.getLoginUserInfo(request);
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
-        if (userSession != null ) {
-            if (b_uidx.equals(userSession.getIdx())) {
-                BoardUpdateRequest boardUpdateRequest = new BoardUpdateRequest(bIdx, boardRequest.getTitle(), boardRequest.getContent(), new Timestamp(new Date().getTime()));
-                System.out.println("boardupdatetitle: "+boardRequest.getTitle());
-                System.out.println("boardupdatecontent: "+boardRequest.getContent());
-                boardMapper.updatePost(boardUpdateRequest);
-                return "게시물 수정 완료";
+        // mybatis에서 본인 idx와 게시물 idx가 같으면
+        BoardResponse board = boardMapper.selectPostByPostId(bIdx);
+
+        if (board != null) {
+            Integer b_uidx = board.getUserIdx();
+
+            if (userSession != null ) {
+                if (b_uidx.equals(userSession.getIdx())) {
+                    Date now = new Date();
+                    BoardUpdateRequest boardUpdateRequest = new BoardUpdateRequest(bIdx, boardRequest.getTitle(), boardRequest.getContent(), new Date());
+                    boardMapper.updatePost(boardUpdateRequest);
+                    return "게시물 수정 완료";
+                } else {
+                    System.out.println("본인 게시물만 삭제 가능");
+                }
             } else {
-                System.out.println("본인 게시물만 삭제 가능");
+                System.out.println("사용자 로그인 정보 없음");
             }
         } else {
-            System.out.println("사용자 로그인 정보 없음");
+            System.out.println("게시물이 존재하지 않음");
         }
         return "게시물 수정 실패";
     }
 
-    public String deletePost(Integer bIdx, HttpServletRequest request) {
+    public String deletePost(Integer bIdx) throws Exception {
 
-        UserSession userSession = userService.getLoginUserInfo(request);
-        Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+        // MyBatis에서 본인의 idx와 idx가 같다면
+//        UserSession userSession = userService.getLoginUserInfo(request);
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
+        BoardResponse board = boardMapper.selectPostByPostId(bIdx);
 
-        if (userSession != null ) {
-            if (b_uidx.equals(userSession.getIdx())) {
-                boardMapper.deletePost(bIdx);
-                return "게시물 삭제 완료";
+        if (board != null) {
+            Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+
+            if (userSession != null ) {
+                if (b_uidx.equals(userSession.getIdx())) {
+                    boardMapper.deletePost(bIdx);
+                    return "게시물 삭제 완료";
+                } else {
+                    System.out.println("본인 게시물만 삭제 가능");
+                }
             } else {
-                System.out.println("본인 게시물만 삭제 가능");
+                System.out.println("사용자 로그인 정보 없음");
             }
         } else {
-            System.out.println("사용자 로그인 정보 없음");
+            System.out.println("게시물이 존재하지 않음");
         }
         return "게시물 삭제 실패";
     }
-
-    public BoardResponse searchPosts(String q) {
-        return boardMapper.searchPosts(q);
-    }
-
-    // 제목 검색 like 0%0
-    // 내용 검색
-    // 작성자 검색
 }
